+++ sdk/sdk-api/src/main/java/mil/emp3/api/global.java
@@ -4,6 +4,7 @@ import mil.emp3.api.interfaces.IContainer;
 import mil.emp3.api.interfaces.IEmpPropertyList;
 import mil.emp3.api.interfaces.IFeature;
 import mil.emp3.api.interfaces.IOverlay;
+import mil.emp3.api.interfaces.core.ICoreManager;
 import mil.emp3.api.interfaces.core.IStorageManager;
 import mil.emp3.api.utils.ManagerFactory;
